requirements.txt
evidently
nltk
pandas
langchain-ollama
ollama
uvicorn
fastapi
huggingface-hub
sentence-transformers
pymongo
python-dotenv
faiss-cpu
langchain-google-genai
black